% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectSkin.R
\docType{class}
\name{TrenaProjectSkin-class}
\alias{TrenaProjectSkin-class}
\alias{.TrenaProjectSkin}
\alias{TrenaProjectSkin}
\title{TrenaProjectSkin-class}
\usage{
TrenaProjectSkin(quiet = TRUE)
}
\value{
An object of the TrenaProjectSkin class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
